+++ waterfun-service-core/src/main/java/org/waterwood/waterfunservicecore/entity/audit/AuditTaskStatusConverter.java
@@ -4,15 +4,15 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
-public class AuditTaskStatusConverter implements AttributeConverter<AuditStatus, Short> {
+public class AuditTaskStatusConverter implements AttributeConverter<AuditStatus, Byte> {
 
     @Override
-    public Short convertToDatabaseColumn(AuditStatus auditTaskStatus) {
+    public Byte convertToDatabaseColumn(AuditStatus auditTaskStatus) {
         return auditTaskStatus == null ? null : auditTaskStatus.getCode();
     }
 
     @Override
-    public AuditStatus convertToEntityAttribute(Short aShort) {
+    public AuditStatus convertToEntityAttribute(Byte aShort) {
         return aShort == null ? null : AuditStatus.fromCode(aShort);
     }
 }
