+++ waterfun-service-core/src/main/java/org/waterwood/waterfunservicecore/entity/notification/BusinessTypeConverter.java
@@ -4,15 +4,15 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
-public class BusinessTypeConverter implements AttributeConverter<BusinessType, Short> {
+public class BusinessTypeConverter implements AttributeConverter<BusinessType, Byte> {
 
     @Override
-    public Short convertToDatabaseColumn(BusinessType attribute) {
+    public Byte convertToDatabaseColumn(BusinessType attribute) {
         return attribute == null ? null : attribute.getValue();
     }
 
     @Override
-    public BusinessType convertToEntityAttribute(Short dbData) {
+    public BusinessType convertToEntityAttribute(Byte dbData) {
         return dbData == null ? null : BusinessType.fromValue(dbData);
     }
 }
