+++ waterfun-service-core/src/main/java/org/waterwood/waterfunservicecore/entity/post/PostEditStatusConvertor.java
@@ -4,15 +4,15 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
-public class PostEditStatusConvertor implements AttributeConverter<PostEditStatus, Short> {
+public class PostEditStatusConvertor implements AttributeConverter<PostEditStatus, Byte> {
 
     @Override
-    public Short convertToDatabaseColumn(PostEditStatus attribute) {
+    public Byte convertToDatabaseColumn(PostEditStatus attribute) {
         return attribute == null ? null : attribute.getValue();
     }
 
     @Override
-    public PostEditStatus convertToEntityAttribute(Short dbData) {
+    public PostEditStatus convertToEntityAttribute(Byte dbData) {
         return PostEditStatus.fromCode(dbData);
     }
 }
