+++ waterfun-service-core/src/main/java/org/waterwood/waterfunservicecore/entity/post/PostStatusConvertor.java
@@ -4,15 +4,15 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
-public class PostStatusConvertor implements AttributeConverter<PostStatus, Short> {
+public class PostStatusConvertor implements AttributeConverter<PostStatus, Byte> {
 
     @Override
-    public Short convertToDatabaseColumn(PostStatus attribute) {
+    public Byte convertToDatabaseColumn(PostStatus attribute) {
         return attribute == null ? 0 : attribute.getValue();
     }
 
     @Override
-    public PostStatus convertToEntityAttribute(Short dbData) {
+    public PostStatus convertToEntityAttribute(Byte dbData) {
         return PostStatus.fromCode(dbData);
     }
 }
